+++ code/game/bg_lib.c
@@ -51,7 +51,7 @@ typedef int		 cmp_t(const void *, const void *);
 #endif
 
 static char* med3(char *, char *, char *, cmp_t *);
-static void	 swapfunc(char *, char *, int, int);
+static void	 swapfunc(char *, char *, size_t, size_t);
 
 #ifndef min
 #define min(a, b)	(a) < (b) ? a : b
@@ -61,7 +61,7 @@ static void	 swapfunc(char *, char *, int, int);
  * Qsort routine from Bentley & McIlroy's "Engineering a Sort Function".
  */
 #define swapcode(TYPE, parmi, parmj, n) { 		\
-	long i = (n) / sizeof (TYPE); 			\
+	size_t i = (n) / sizeof (TYPE); 			\
 	register TYPE *pi = (TYPE *) (parmi); 		\
 	register TYPE *pj = (TYPE *) (parmj); 		\
 	do { 						\
@@ -77,7 +77,7 @@ static void	 swapfunc(char *, char *, int, int);
 static void
 swapfunc(a, b, n, swaptype)
 	char *a, *b;
-	int n, swaptype;
+	size_t n, swaptype;
 {
 	if(swaptype <= 1)
 		swapcode(long, a, b, n)
@@ -87,9 +87,9 @@ swapfunc(a, b, n, swaptype)
 
 #define swap(a, b)					\
 	if (swaptype == 0) {				\
-		long t = *(long *)(a);			\
-		*(long *)(a) = *(long *)(b);		\
-		*(long *)(b) = t;			\
+		size_t t = *(size_t *)(a);			\
+		*(size_t *)(a) = *(size_t *)(b);		\
+		*(size_t *)(b) = t;			\
 	} else						\
 		swapfunc(a, b, es, swaptype)
 
@@ -112,7 +112,7 @@ qsort(a, n, es, cmp)
 	cmp_t *cmp;
 {
 	char *pa, *pb, *pc, *pd, *pl, *pm, *pn;
-	int d, r, swaptype, swap_cnt;
+	size_t d, r, swaptype, swap_cnt;
 
 loop:	SWAPINIT(a, es);
 	swap_cnt = 0;
